+++ test_resources/android/build_flavorizr_processor_test/flavorizr_expected.gradle
@@ -1,0 +1,29 @@
+android {
+    flavorDimensions += "flavor-type"
+
+    productFlavors {
+        apple {
+            dimension "flavor-type"
+            applicationId "com.example.apple"
+            versionNameSuffix "-green-prod"
+            signingConfig flavorSigning.green
+            versionCode 1000
+            minSdkVersion 23
+            resValue "string", "app_name", "Apple App"
+            resValue "string", "variable_one", "previous variable one"
+            resValue "string", "common", "test common"
+            buildConfigField "String", "field_one", "\"previous field one\""
+        }
+        banana {
+            dimension "flavor-type"
+            applicationId "com.example.banana"
+            resValue "string", "app_name", "Banana App"
+            resValue "string", "variable_one", "test variable one"
+            resValue "string", "common", "test common"
+            resValue "string", "variable_two", "test variable two"
+            buildConfigField "String", "field_one", "\"test field one\""
+            buildConfigField "char", "field_two", '\'y\''
+            buildConfigField "double", "field_three", "20.0"
+        }
+    }
+}
