+++ test_resources/android/build_gradle_processor_test/build_expected.gradle
@@ -28,29 +28,6 @@ apply plugin: 'kotlin-android'
 apply from: "$flutterRoot/packages/flutter_tools/gradle/flutter.gradle"
 
 android {
-
-    // ----- BEGIN flavorDimensions (autogenerated by flutter_flavorizr) -----
-    flavorDimensions += "flavor-type"
-
-    productFlavors {
-        apple {
-            dimension "flavor-type"
-            applicationId "com.example.apple"
-            resValue "string", "app_name", "Apple App"
-        }
-        banana {
-            dimension "flavor-type"
-            applicationId "com.example.banana"
-            resValue "string", "app_name", "Banana App"
-            resValue "string", "variable_one", "test variable one"
-            resValue "string", "variable_two", "test variable two"
-            buildConfigField "String", "field_one", "\"test field one\""
-            buildConfigField "char", "field_two", '\'y\''
-            buildConfigField "double", "field_three", "20.0"
-        }
-    }
-    // ----- END flavorDimensions (autogenerated by flutter_flavorizr) -----
-
     compileSdkVersion 28
 
     sourceSets {
@@ -91,3 +68,6 @@ dependencies {
     androidTestImplementation 'androidx.test.espresso:espresso-core:3.5.1'
 }
 
+// ----- BEGIN flavorDimensions (autogenerated by flutter_flavorizr) -----
+apply from: "flavorizr.gradle"
+// ----- END flavorDimensions (autogenerated by flutter_flavorizr) -----
