+++ test_resources/android/build_gradle_processor_test/build_idempotent_1.gradle
