+++ test_resources/android/build_gradle_processor_test/build_idempotent_2.gradle
@@ -67,3 +67,7 @@ dependencies {
     androidTestImplementation 'androidx.test:runner:1.5.2'
     androidTestImplementation 'androidx.test.espresso:espresso-core:3.5.1'
 }
+
+// ----- BEGIN flavorDimensions (autogenerated by flutter_flavorizr) -----
+apply from: "flavorizr.gradle"
+// ----- END flavorDimensions (autogenerated by flutter_flavorizr) -----
