+++ test_resources/android/build_gradle_processor_test/build_original.gradle
