--- test_resources/android/build_gradle_processor_with_custom_config_test/build_expected.gradle
@@ -1,101 +1,0 @@
-package test_resources.android.build_gradle_processor_test
-
-def localProperties = new Properties()
-def localPropertiesFile = rootProject.file('local.properties')
-if (localPropertiesFile.exists()) {
-    localPropertiesFile.withReader('UTF-8') { reader ->
-        localProperties.load(reader)
-    }
-}
-
-def flutterRoot = localProperties.getProperty('flutter.sdk')
-if (flutterRoot == null) {
-    throw new GradleException("Flutter SDK not found. Define location with flutter.sdk in the local.properties file.")
-}
-
-def flutterVersionCode = localProperties.getProperty('flutter.versionCode')
-if (flutterVersionCode == null) {
-    flutterVersionCode = '1'
-}
-
-def flutterVersionName = localProperties.getProperty('flutter.versionName')
-if (flutterVersionName == null) {
-    flutterVersionName = '1.0'
-}
-
-apply plugin: 'com.android.application'
-apply plugin: 'kotlin-android'
-apply from: "$flutterRoot/packages/flutter_tools/gradle/flutter.gradle"
-
-android {
-
-    // ----- BEGIN flavorDimensions (autogenerated by flutter_flavorizr) -----
-    flavorDimensions += "flavor-type"
-
-    productFlavors {
-        apple {
-            dimension "flavor-type"
-            applicationId "com.example.apple"
-            versionNameSuffix "-green-prod"
-            signingConfig flavorSigning.green
-            versionCode 1000
-            minSdkVersion 23
-            resValue "string", "app_name", "Apple App"
-            resValue "string", "variable_one", "previous variable one"
-            resValue "string", "common", "test common"
-            buildConfigField "String", "field_one", "\"previous field one\""
-        }
-        banana {
-            dimension "flavor-type"
-            applicationId "com.example.banana"
-            resValue "string", "app_name", "Banana App"
-            resValue "string", "variable_one", "test variable one"
-            resValue "string", "common", "test common"
-            resValue "string", "variable_two", "test variable two"
-            buildConfigField "String", "field_one", "\"test field one\""
-            buildConfigField "char", "field_two", '\'y\''
-            buildConfigField "double", "field_three", "20.0"
-        }
-    }
-    // ----- END flavorDimensions (autogenerated by flutter_flavorizr) -----
-
-    compileSdkVersion 28
-
-    sourceSets {
-        main.java.srcDirs += 'src/main/kotlin'
-    }
-
-    lintOptions {
-        disable 'InvalidPackage'
-    }
-
-    defaultConfig {
-        // TODO: Specify your own unique Application ID (https://developer.android.com/studio/build/application-id.html).
-        applicationId "com.mylittlesuite.flutterflavorizr_example"
-        minSdkVersion 16
-        targetSdkVersion 28
-        versionCode flutterVersionCode.toInteger()
-        versionName flutterVersionName
-        testInstrumentationRunner "androidx.test.runner.AndroidJUnitRunner"
-    }
-
-    buildTypes {
-        release {
-            // TODO: Add your own signing config for the release build.
-            // Signing with the debug keys for now, so `flutter run --release` works.
-            signingConfig signingConfigs.debug
-        }
-    }
-}
-
-flutter {
-    source '../..'
-}
-
-dependencies {
-    implementation "org.jetbrains.kotlin:kotlin-stdlib-jdk7:$kotlin_version"
-    testImplementation 'junit:junit:4.13.2'
-    androidTestImplementation 'androidx.test:runner:1.5.2'
-    androidTestImplementation 'androidx.test.espresso:espresso-core:3.5.1'
-}
-
